heap basic operations

+++ src/test/test_helper.cpp
@@ -1,8 +1,7 @@
 #include "test_helper.h"
 
-vector<int> createVector(int size) {
+vector<int> createVector(size_t size) {
 	vector<int> vect(size);
 	generate(vect.begin(), vect.end(), Generator<int>(size));
 	return vect;
 }
-
